+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+
+## master
+
+* Add support for pulling API credentials & symbols from environmental
+  variables and a configuration file.
+
+
 ## v0.1.0.0
 
 * Initial release